requirements.txt
Flask
google-generativeai
Gunicorn